+++ x2engine/js/lib/aight/aight.js
@@ -527,31 +527,31 @@ if (!Array.isArray) {
 var boxedString = Object("a"),
     splitString = boxedString[0] != "a" || !(0 in boxedString);
 
-if (!Array.prototype.forEach) {
-    Array.prototype.forEach = function forEach(fun /*, thisp*/) {
-        var object = toObject(this),
-            self = splitString && _toString(this) == "[object String]" ?
-                this.split("") :
-                object,
-            thisp = arguments[1],
-            i = -1,
-            length = self.length >>> 0;
-
-        // If no callback function or if callback is not a callable function
-        if (_toString(fun) != "[object Function]") {
-            throw new TypeError(); // TODO message
-        }
-
-        while (++i < length) {
-            if (i in self) {
-                // Invoke the callback function with call, passing arguments:
-                // context, property value, property key, thisArg object
-                // context
-                fun.call(thisp, self[i], i, object);
-            }
-        }
-    };
-}
+//if (!Array.prototype.forEach) {
+//    Array.prototype.forEach = function forEach(fun /*, thisp*/) {
+//        var object = toObject(this),
+//            self = splitString && _toString(this) == "[object String]" ?
+//                this.split("") :
+//                object,
+//            thisp = arguments[1],
+//            i = -1,
+//            length = self.length >>> 0;
+//
+//        // If no callback function or if callback is not a callable function
+//        if (_toString(fun) != "[object Function]") {
+//            throw new TypeError(); // TODO message
+//        }
+//
+//        while (++i < length) {
+//            if (i in self) {
+//                // Invoke the callback function with call, passing arguments:
+//                // context, property value, property key, thisArg object
+//                // context
+//                fun.call(thisp, self[i], i, object);
+//            }
+//        }
+//    };
+//}
 
 // ES5 15.4.4.19
 // http://es5.github.com/#x15.4.4.19