+++ x2engine/protected/config/X2Config.php
@@ -5,8 +5,8 @@ $host = 'localhost';
 $user = 'root';
 $pass = '';
 $dbname = 'x2engine';
-$version = '5.0';
-$buildDate = '1417848242';
-$updaterVersion = '5.1';
+$version = '5.0.1';
+$buildDate = '1418174564';
+$updaterVersion = '5.1.1';
 $language='en';
 ?>